requirements.txt
fastapi==0.115.12
librosa==0.11.0
numpy==2.2.6
Pillow==11.2.1
pydantic==2.11.4
starlette==0.46.2
uvicorn==0.34.2
